Pipes propeties now node properties with expanded props
